moved database functions from database.py to dbfunctions.py
