+++ dbfunctions.py
@@ -1,0 +1,26 @@
+from database import Account
+import auth
+
+# instantiating table objects
+student_account = Account("meow.db")
+student_account_backup = Account("backup.db")
+
+def create_account(username: str, password: str):
+    # checks for valid username and password is already done 
+    # check for repeated username
+    if student_account.retrieve("username", username) is None:
+        password, salt = auth.create_hash(password)
+        student_account.insert(username, password, salt)
+        student_account_backup.insert(username, password, salt)
+
+def login(username: str , password: str) -> bool:
+    # checks for valid username and password is already done 
+
+    data = student_account.retrieve("username", username)
+    # account not found
+    if data is None:
+        return False
+
+    account_id, database_username, database_password, database_salt = data
+    # salting and hashing of password implemented 
+    return auth.check_password(password, database_password, database_salt)
